requirements.txt
finnhub_python==2.4.7
requests==2.22.0
termcolor==1.1.0
